install supertest and refactor app.js file
