+++ package.json
@@ -14,10 +14,14 @@
   },
   "dependencies": {
     "express": "^4.17.1",
-    "mongoose": "^5.10.7"
+    "mongoose": "^5.10.7",
+    "supertest": "^5.0.0"
   },
   "devDependencies": {
     "eslint-config-hackreactor": "git://github.com/reactorcore/eslint-config-hackreactor",
     "jest": "^26.4.2"
+  },
+  "jest": {
+    "testEnvironment": "node"
   }
 }
