+++ server/index.js
@@ -1,12 +1,4 @@
-const express = require("express");
-const app = express();
-
-const db = require("../database/index.js");
-
-app.get("/photos/:productid/:styleid", async (req, res) => {
-  const photos = await db.getPhotosById(req.params.productid, req.params.styleid);
-  res.send(photos);
-})
+const app = require("./app");
 
 const port = 3000;
