+++ server/index.test.js
@@ -1,0 +1,5 @@
+describe("Photo API", () => {
+  test("placeholder", () => {
+    expect(1+1).toBe(2);
+  })
+})
